Move arraySetDifference to lib
