+++ apps/spruce/src/utils/array/array.test.ts
@@ -1,6 +1,5 @@
 import {
   arrayIntersection,
-  arraySetDifference,
   arrayUnion,
   convertArrayToObject,
   convertObjectToArray,
@@ -198,40 +197,6 @@ describe("arrayIntersection", () => {
     expect(
       arrayIntersection(["1", "1", "2", "2", "3"], ["2", "3", "4"]),
     ).toStrictEqual(["2", "3"]);
-  });
-});
-
-describe("arraySetDifference", () => {
-  it("should throw an error if an object is passed in", () => {
-    expect(() => arraySetDifference([{}], [{}])).toThrow(
-      TypeError("arraySetDifference does not support objects"),
-    );
-  });
-  it("should return an empty array when the arrays are empty", () => {
-    expect(arraySetDifference([], [])).toStrictEqual([]);
-  });
-  it("should return the differing values when the first array is empty", () => {
-    expect(arraySetDifference([], ["1", "2", "3"])).toStrictEqual([]);
-  });
-  it("should return the differing values when the second array is empty", () => {
-    expect(arraySetDifference(["1", "2", "3"], [])).toStrictEqual([
-      "1",
-      "2",
-      "3",
-    ]);
-  });
-  it("should return the differing values when the arrays have no common elements", () => {
-    expect(arraySetDifference(["1", "2", "3"], ["4", "5", "6"])).toStrictEqual([
-      "1",
-      "2",
-      "3",
-    ]);
-  });
-  it("should return the differing values when the arrays have common elements", () => {
-    expect(arraySetDifference(["1", "2", "3"], ["3", "4", "5"])).toStrictEqual([
-      "1",
-      "2",
-    ]);
   });
 });
