+++ apps/spruce/src/utils/array/index.ts
@@ -168,23 +168,6 @@ export const arrayIntersection = <T>(a: T[], b: T[]) => {
   return intersection;
 };
 
-/**
- * `arraySetDifference` returns the elements in a that are not in b
- * @param a - The first array
- * @param b - The second array
- * @returns The elements in a that are not in b
- * @example arraySetDifference([1, 2, 3], [2, 3, 4]) // [1]
- */
-export const arraySetDifference = <T>(a: T[], b: T[]) => {
-  if (typeof a[0] === "object" || typeof b[0] === "object") {
-    throw new TypeError("arraySetDifference does not support objects");
-  }
-  const setA = new Set(a);
-  const setB = new Set(b);
-  const difference = Array.from(setA).filter((x) => !setB.has(x));
-  return difference;
-};
-
 type SortFunction<T> = (a: T, b: T) => number;
 
 /**
